+++ assets/css/app.css
@@ -44,3 +44,8 @@ html[data-theme="dark"] {
         @apply border-l-gray-800;
     }
 }
+
+.validate-error {
+    margin-top: calc(0.25rem * 2);
+    font-size: 0.75rem;
+}
