+++ package.json
@@ -24,11 +24,12 @@
         "@pinia/nuxt": "0.11.1",
         "@tailwindcss/vite": "^4.1.11",
         "@vee-validate/nuxt": "^4.15.1",
+        "@vee-validate/rules": "^4.15.1",
         "@vee-validate/zod": "^4.15.1",
         "better-auth": "^1.2.12",
         "concurrently": "^9.2.0",
-        "drizzle-orm": "^0.44.2",
-        "drizzle-zod": "^0.8.2",
+        "drizzle-orm": "^0.40.1",
+        "drizzle-zod": "^0.7.0",
         "maplibre-gl": "^5.6.1",
         "nanoid": "^5.1.5",
         "nuxt": "^3.17.6",
@@ -36,9 +37,10 @@
         "nuxt-maplibre": "1.1.1",
         "pinia": "^3.0.3",
         "tailwindcss": "^4.1.11",
+        "vee-validate": "4.15.1",
         "vue": "^3.5.17",
         "vue-router": "^4.5.1",
-        "zod": "^3.25.76"
+        "zod": "^3.24.2"
     },
     "devDependencies": {
         "@antfu/eslint-config": "^4.16.2",
